requirements.txt
viam-sdk
smbus
